1.3

+++ core/src/main/java/fr/formiko/kokcinelo/view/MapActor.java
@@ -147,8 +147,8 @@ public class MapActor extends Actor {
         final float egdeSize = 0.2f;
 
         Pixmap darkedArea = new Pixmap((int) getWidth(), (int) getHeight(), Pixmap.Format.RGBA8888);
-        int xCenter = (int) (darkedArea.getWidth() / 2);
-        int yCenter = (int) (darkedArea.getHeight() / 2);
+        int xCenter = darkedArea.getWidth() / 2;
+        int yCenter = darkedArea.getHeight() / 2;
         float edgeLength = radius * egdeSize;
 
         for (int x = 0; x < darkedArea.getWidth(); x++) {
